试剂打印方式修改，增加日志，细节调整

+++ yudao-module-reagent/src/main/java/cn/iocoder/yudao/module/reagent/service/ReagentBaseFlatServiceImpl.java
@@ -7,9 +7,16 @@ import cn.iocoder.yudao.module.reagent.controller.admin.vo.ReagentBaseFlatSaveRe
 import cn.iocoder.yudao.module.reagent.controller.admin.vo.ReagentBaseFlatSimplePageReqVO;
 import cn.iocoder.yudao.module.reagent.dal.dataobject.ReagentBaseFlatDO;
 import cn.iocoder.yudao.module.reagent.dal.mysql.ReagentBaseFlatMapper;
+import com.mzt.logapi.context.LogRecordContext;
+import com.mzt.logapi.service.impl.DiffParseFunction;
+import com.mzt.logapi.starter.annotation.LogRecord;
 import jakarta.annotation.Resource;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+
+import static cn.iocoder.yudao.framework.common.exception.util.ServiceExceptionUtil.exception;
+import static cn.iocoder.yudao.module.reagent.enums.ErrorCodeConstants.REAGENT_BASE_NOT_EXISTS;
+import static cn.iocoder.yudao.module.reagent.enums.LogRecordConstants.*;
 
 /**
  * 老ERP同步 试剂基础数据(扁平) Service 实现
@@ -29,12 +36,21 @@ public class ReagentBaseFlatServiceImpl implements ReagentBaseFlatService {
     }
 
     @Override
+    @LogRecord(type = REAGENT_BASE_FLAT_TYPE, subType = REAGENT_BASE_FLAT_UPDATE_SUB_TYPE, bizNo = "{{#updateReqVO.id}}",
+            success = REAGENT_BASE_FLAT_UPDATE_SUCCESS)
     public void updateReagentBaseFlat(ReagentBaseFlatSaveReqVO updateReqVO) {
+        ReagentBaseFlatDO oldBaseFlat = reagentBaseFlatMapper.selectById(updateReqVO.getId());
+        if (oldBaseFlat == null) {
+            throw exception(REAGENT_BASE_NOT_EXISTS);
+        }
         // 只更新非空可维护字段（MyBatis-Plus 默认 NOT_NULL 策略，同步来源字段不会被动）
         ReagentBaseFlatDO updateObj = BeanUtils.toBean(updateReqVO, ReagentBaseFlatDO.class);
         reagentBaseFlatMapper.updateById(updateObj);
         log.info("[reagent-base-flat] 更新 {}：名称={}, 供应商={}, 货号={}", updateReqVO.getId(),
                 updateReqVO.getReagentName(), updateReqVO.getVendor(), updateReqVO.getCatNo());
+        LogRecordContext.putVariable("baseFlat", oldBaseFlat);
+        LogRecordContext.putVariable(DiffParseFunction.OLD_OBJECT,
+                BeanUtils.toBean(oldBaseFlat, ReagentBaseFlatSaveReqVO.class));
     }
 
     @Override
@@ -43,9 +59,16 @@ public class ReagentBaseFlatServiceImpl implements ReagentBaseFlatService {
     }
 
     @Override
+    @LogRecord(type = REAGENT_BASE_FLAT_TYPE, subType = REAGENT_BASE_FLAT_DELETE_SUB_TYPE, bizNo = "{{#id}}",
+            success = REAGENT_BASE_FLAT_DELETE_SUCCESS)
     public void deleteReagentBaseFlat(Long id) {
+        ReagentBaseFlatDO baseFlat = reagentBaseFlatMapper.selectById(id);
+        if (baseFlat == null) {
+            throw exception(REAGENT_BASE_NOT_EXISTS);
+        }
         reagentBaseFlatMapper.deleteByIdPhysical(id);
         log.info("[reagent-base-flat] 删除 {}：id={}", id);
+        LogRecordContext.putVariable("baseFlat", baseFlat);
     }
 
 }
